Change long to int for Python 3
We do not need long here

+++ src/twisted/words/protocols/irc.py
@@ -2702,7 +2702,7 @@ def dccParseAddress(address):
         pass
     else:
         try:
-            address = long(address)
+            address = int(address)
         except ValueError:
             raise IRCBadMessage("Indecipherable address %r" % (address,))
         else:
@@ -2978,7 +2978,7 @@ def dccDescribe(data):
         pass
     else:
         try:
-            address = long(address)
+            address = int(address)
         except ValueError:
             pass
         else:
